requirements.txt
# Core OpenAI and API Dependencies
openai>=1.3.0  # OpenAI API client
python-dotenv>=0.19.0  # For environment variables

# Terminal UI Dependencies
rich>=13.0.0  # For rich terminal interface and formatting

# Utility Dependencies
tenacity>=8.0.0  # For retry functionality
typing-extensions>=4.0.0  # For type hints and overrides

# Gmail API Dependencies
google-auth-oauthlib>=1.0.0  # For Gmail OAuth2
google-auth-httplib2>=0.1.0  # HTTP client for Google APIs
google-api-python-client>=2.0.0  # Google API client library
